+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: lhelper <lhelper@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/10 17:53:17 by mkristie          #+#    #+#             */
-/*   Updated: 2020/10/23 18:39:24 by lhelper          ###   ########.fr       */
+/*   Updated: 2020/10/24 16:13:06 by lhelper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ extern _Bool				g_squote;
 extern _Bool				g_backslash;
 extern int					g_backslash_time;
 extern int					g_isfork;
-extern int					g_exit;
+extern long long			g_exit;
 
 void	handle_cmd(char **cmd);//
 char **list_to_env();//
